add TestPostUserBalanceWithdraw

+++ internal/luhn/lunh.go
@@ -1,11 +1,11 @@
 package luhn
 
-func Valid(number int) bool {
+func Valid(number int64) bool {
 	return (number%10+checksum(number/10))%10 == 0
 }
 
-func checksum(number int) int {
-	var luhn int
+func checksum(number int64) int64 {
+	var luhn int64
 
 	for i := 0; number > 0; i++ {
 		cur := number % 10
